+++ CHANGELOG.md
@@ -3,16 +3,36 @@ Polyverse Boost Cloud Service Monitor
 
 # Release Notes
 
-## Version 0.2.0: July 1st, 2023
+## Version 0.3.0: July 2nd, 2023
+
+### New Features
+- Service Check on Flow Diagram
+- Service Check on Blueprint
+- Service Check on Convert
+- Service Check on Test Gen
+- Service Check on Compliance
+- Service Check on Compliance Function
+- Service Check on Analyze
+- Service Check on Analyze Function
+- Service Check on Summarize
+- Service Check on Custom
+
+### Enhancements
+- N/A
+
+### Bug Fixes
+- N/A
+
+## Version 0.2.0: July 2nd, 2023
 
 ### New Features
 - Service Check on Explain
 
-# Enhancements
+### Enhancements
 - Functional tests for service check responses
 - Blending of input data for service checks
 
-# Bug Fixes
+### Bug Fixes
 - N/A
 
 ## Version 0.1.0: July 1st, 2023
@@ -21,8 +41,8 @@ Polyverse Boost Cloud Service Monitor
 - Service Check on User Organizations
 - Service Check on Customer Portal
 
-# Enhancements
+### Enhancements
 - N/A
 
-# Bug Fixes
+### Bug Fixes
 - N/A